track page type snippet dialog
