+++ packages/slice-machine/src/features/customTypes/customTypesBuilder/PageSnippetDialog/PageSnippetDialog.tsx
@@ -13,11 +13,15 @@ import { ContentTabs } from "@src/components/ContentTabs";
 import { MarkdownRenderer } from "@src/features/documentation/MarkdownRenderer";
 import { useDocumentation } from "@src/features/documentation/useDocumentation";
 
+import { useSliceMachineConfig } from "@src/hooks/useSliceMachineConfig";
+
 import * as styles from "./PageSnippetDialog.css";
+import { telemetry } from "@src/apiClient";
 
 type PageSnippetContentProps = { model: CustomType };
 
 const PageSnippetContent: FC<PageSnippetContentProps> = ({ model }) => {
+  const config = useSliceMachineConfig();
   const documentation = useDocumentation({
     kind: "PageSnippet",
     data: { model },
@@ -27,11 +31,22 @@ const PageSnippetContent: FC<PageSnippetContentProps> = ({ model }) => {
     return null;
   }
 
+  const trackOpenSnippet = () => {
+    void telemetry.track({
+      event: "page-type:open-snippet",
+      framework: config?.adapter,
+    });
+  };
+
   return (
     <Dialog
       size="small"
       trigger={
-        <Button variant="secondary" startIcon={<Icon name="code" />}>
+        <Button
+          onClick={() => trackOpenSnippet()}
+          variant="secondary"
+          startIcon={<Icon name="code" />}
+        >
           Page snippet
         </Button>
       }
